updates and modernizes

+++ install.sh
@@ -1,25 +1,20 @@
 #!/usr/bin/env zsh
-############################
-# This script creates symlinks from the home directory to any desired dotfiles in $HOME/dotfiles
-# And also installs MacOS Software
-# And also installs Homebrew Packages and Casks (Apps)
-# And also sets up VS Code
-############################
+set -e
 
-# dotfiles directory
-dotfiledir="${HOME}/dotfiles"
-
-# list of files/folders to symlink in ${homedir}
+dotfiledir="$(cd "$(dirname "$0")" && pwd)"
 files=(zshrc zprofile zprompt bashrc bash_profile bash_prompt aliases private)
 
-# change to the dotfiles directory
 echo "Changing to the ${dotfiledir} directory"
 cd "${dotfiledir}" || exit
 
 # create symlinks (will overwrite old dotfiles)
 for file in "${files[@]}"; do
-    echo "Creating symlink to $file in home directory."
-    ln -sf "${dotfiledir}/.${file}" "${HOME}/.${file}"
+    if [ -f "${dotfiledir}/.${file}" ]; then
+        echo "Creating symlink to $file in home directory."
+        ln -sf "${dotfiledir}/.${file}" "${HOME}/.${file}"
+    else
+        echo "Warning: ${dotfiledir}/.${file} does not exist. Skipping..."
+    fi
 done
 
 # Run the MacOS Script
